Clean conversion warnings and use size_t length arguments

+++ src/compiler/catalog.h
@@ -96,14 +96,14 @@ static void install_symbol(void *context, fb_symbol_t *sym)
     if (catalog->qualify_names) {
         scope_name = ct->scope->name;
         while (scope_name) {
-            n = scope_name->ident->len;
+            n = (int)scope_name->ident->len;
             memcpy(s, scope_name->ident->text, (size_t)n);
             s += n;
             *s++ = '.';
             scope_name = scope_name->link;
         }
     }
-    n = sym->ident->len;
+    n = (int)sym->ident->len;
     memcpy(s, sym->ident->text, (size_t)n);
     s += n;
     *s++ = '\0';
